remove TMPDIR fiddling to get tmux to work predictably
Apparently for some persistence needed by the sorin theme there was a mod
made to the zprofile to generate a random TMPDIR so each zsh session would have
its own sandboxed tmp directory.

However tmux relies on the tmp dir to be the same for all sessions of the same
user on a system so this wreaked havoc with trying to attach to sessions. Tmux
would always report that there were no sessions available even when *ps ax*
clearly shows several tmux processes running active. The trick to send signal 10
to the tmux process to recreate the socket did not work as the folder where it
was supposed to write it to was missing. In any case this would have lead to
folder and file leaks in the tmp folder which is a finite resource.

So 'raus damit' and back to vanilla TMPDIR.